+++ bioengine/__init__.py
@@ -1,0 +1,5 @@
+"""BioEngine - Engine for the model zoo."""
+
+__version__ = "0.1.0"
+__author__ = "BioEngine Authors <example@email.com>"
+__all__ = []
