--- ioengine/__init__.py
@@ -1,5 +1,0 @@
-"""IOEngine - Engine for the model zoo."""
-
-__version__ = "0.1.0"
-__author__ = "IOEngine Authors <example@email.com>"
-__all__ = []
