+++ tests/__init__.py
@@ -1,1 +1,1 @@
-"""Tests for the ioengine."""
+"""Tests for the bioengine."""
