+++ tox.ini
@@ -17,8 +17,8 @@ basepython = python3
 ignore_errors = True
 commands =
   black --check ./
-  flake8 ioengine tests
-  pylint ioengine tests
-  pydocstyle ioengine tests
+  flake8 bioengine tests
+  pylint bioengine tests
+  pydocstyle bioengine tests
 deps =
   -rrequirements_lint.txt
